CMake: Added contrib/blast.
