+++ contrib/blast/blast.h
@@ -21,6 +21,7 @@
   Mark Adler    madler@alumni.caltech.edu
  */
 
+#define local static            /* for local function definitions */
 
 /*
  * blast() decompresses the PKWare Data Compression Library (DCL) compressed
